update file structure, move der sign to unit test data file
